MM-55724: recovery handler for background tasks
Rework the recovery package to allow use other callbacks, such as the
sync job and metrics jobs managed by the cluster scheduler.

Fixes: https://mattermost.atlassian.net/browse/MM-55724
